+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - request auth token on every retry, not just after 403
 - update to use nopoll v 1.0.2
 - provide signal handlers so we shut down properly when INCLUDE_BREAKPAD active
+- send status code and reason in close message 
 
 ## [1.0.2] - 2019-02-08
 - Refactored connection.c and updated corresponding unit tests